requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
dnspython==2.7.0
filelock==3.18.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pipenv==2024.4.1
platformdirs==4.3.7
PyJWT==2.10.1
pymongo==4.11.3
python-dotenv==1.1.0
requests==2.32.3
setuptools==78.1.0
urllib3==2.3.0
virtualenv==20.30.0
Werkzeug==3.1.3
